+++ app/__init__.py
@@ -1,31 +1,18 @@
-from flask import Flask, abort, render_template, session, redirect, url_for, request
-
-from app.crud import create_user, user_list, last_news, create_news
-from app.forms import LoginForm, RegistrationForm
-
+from flask import Flask, abort, render_template, session, request
+from app.crud import last_news, create_news
+from app.user.views import blueprint as user_blueprint
 import os
 
 
 def create_app():
     app = Flask(__name__)
-    app.config.from_pyfile('config.py')
+    app.config.from_pyfile("config.py")
     app.secret_key = os.urandom(32)
+    app.register_blueprint(user_blueprint)
 
     @app.route("/")
     def index():
         return render_template("home.html")
-
-    @app.route('/login')
-    def login():
-        title = "Вход"
-        login_form = LoginForm()
-        return render_template('login.html', page_title=title, form=login_form)
-
-    @app.route("/signup")
-    def registration():
-        title = "Регистрация"
-        reg_form = RegistrationForm()
-        return render_template('registration.html', page_title=title, form=reg_form)
 
     @app.route("/profile/<username>")
     def profile(username):
