+++ app/templates/base.html
@@ -18,40 +18,7 @@
 </head>
 
 <body>
-    <nav class="navbar navbar-dark navbar-expand-lg" style="background-color: #344850;">
-        <div class="container-fluid">
-            <a class="navbar-brand" href="#">
-                <img src="../static/img/LOGO2.png" alt="Логотип приложения" width="45" height="35">
-            </a>
-            <button class="navbar-toggler" type="button" data-bs-toggle=""
-                data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false"
-                aria-label="Toggle navigation">
-                <span class="navbar-toggler-icon"></span>
-            </button>
-            <div class="collapse navbar-collapse" id="navbarSupportedContent">
-                <ul class="navbar-nav me-auto mb-2 mb-lg-0 top-menu">
-                    <li class="nav-item">
-                        <a class="nav-link" href="/">Home</a>
-                    </li>
-                    <li class="nav-item">
-                        <a class="nav-link" href="#">News</a>
-                    </li>
-                </ul>
-
-                <ul class="navbar-nav ms-auto mb-2 mb-lg-0">
-                    <li class="nav-item dropdown">
-                        <a class="nav-link dropdown-toggle" data-bs-toggle="dropdown" href="#"><i
-                                class="far fa-user"></i></a>
-                        <ul class="dropdown-menu" aria-labelledby="navbarDropdown">
-                            <li><a class="dropdown-item" href="signup">Регистрация</a></li>
-                            <li><a class="dropdown-item" href="login">Авторизация</a></li>
-                        </ul>
-                    </li>
-                </ul>
-            </div>
-        </div>
-    </nav>
-
+    {% include 'menu.html' %}
     <section>
         <div class="container">
             <div class="row">
@@ -74,34 +41,10 @@
             </div>
         </div>
     </section>
-
-    <footer class="footer">
-        <div class="container">
-            <div class="row text-md-center">
-                <div class="col-lg-6">
-                    <h4>Информация</h4>
-                    <ul class="list-unstyled">
-                        <li><a href="/">Главная</a></li>
-                        <li><a href="#">Форма обратной связи</a></li>
-                    </ul>
-                </div>
-
-                <div class="col-lg-6">
-                    <h4>Мы в Соц.сетях</h4>
-                    <div class="footer-icons">
-                        <a href="#"><i class="fab fa-facebook-f"></i></a>
-                        <a href="#"><i class="fab fa-vk"></i></a>
-                        <a href="#"><i class="fab fa-instagram"></i></a>
-                    </div>
-                </div>
-            </div>
-        </div>
-
-    </footer>
+    {% include 'footer.html' %}
 
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"
         integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL"
         crossorigin="anonymous"></script>
 </body>
-
 </html>
