+++ app/templates/footer.html
@@ -1,0 +1,21 @@
+<footer class="footer">
+    <div class="container">
+        <div class="row text-md-center">
+            <div class="col-lg-6">
+                <h4>Информация</h4>
+                <ul class="list-unstyled">
+                    <li><a href="/">Главная</a></li>
+                    <li><a href="#">Форма обратной связи</a></li>
+                </ul>
+            </div>
+            <div class="col-lg-6">
+                <h4>Мы в Соц.сетях</h4>
+                <div class="footer-icons">
+                    <a href="#"><i class="fab fa-facebook-f"></i></a>
+                    <a href="#"><i class="fab fa-vk"></i></a>
+                    <a href="#"><i class="fab fa-instagram"></i></a>
+                </div>
+            </div>
+        </div>
+    </div>
+</footer>
