+++ app/templates/menu.html
@@ -1,0 +1,33 @@
+<nav class="navbar navbar-dark navbar-expand-lg" style="background-color: #344850;">
+    <div class="container-fluid">
+        <a class="navbar-brand" href="#">
+            <img src="img/LOGO2.png" alt="Тут логотип" width="45">
+        </a>
+        <button class="navbar-toggler" type="button" data-bs-toggle=""
+            data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false"
+            aria-label="Toggle navigation">
+            <span class="navbar-toggler-icon"></span>
+        </button>
+        <div class="collapse navbar-collapse" id="navbarSupportedContent">
+            <ul class="navbar-nav me-auto mb-2 mb-lg-0 top-menu">
+                <li class="nav-item">
+                    <a class="nav-link" href="/">Домашняя</a>
+                </li>
+                <li class="nav-item">
+                    <a class="nav-link" href="news/index.html">Новости</a>
+                </li>
+            </ul>
+
+            <ul class="navbar-nav ms-auto mb-2 mb-lg-0">
+                <li class="nav-item dropdown">
+                    <a class="nav-link dropdown-toggle" data-bs-toggle="dropdown" href="#"><i
+                            class="far fa-user"></i></a>
+                    <ul class="dropdown-menu" aria-labelledby="navbarDropdown">
+                        <li><a class="dropdown-item" href="{{ url_for('user.registration') }}">Регистрация</a></li>
+                        <li><a class="dropdown-item" href="{{ url_for('user.login') }}">Авторизация</a></li>
+                    </ul>
+                </li>
+            </ul>
+        </div>
+    </div>
+</nav>
