+++ app/templates/user/login.html
@@ -28,6 +28,12 @@ Login page
                     {{ form.password.label }}
                     {{ form.password() }}
                 </div>
+                <div class="form-group form-check p-0">
+                    <h6>
+                        {{ form.remember_me() }}
+                        {{ form.remember_me.label(class_='form-check-label') }}
+                    </h6>
+                </div>
                 {{ form.submit }}
             </form>
         </div>
